+++ docs/source/changelog.rst
@@ -245,3 +245,18 @@ Changelog
 - call `UseExceptions()` on all imported `osgeo` submodules
 - :func:`spatialist.raster.rasterize`: call :meth:`osgeo.gdal.Dataset.FlushCache` to avoid running into suppressed
   errors when closing the dataset with `target_ds = None`
+
+0.17.0 | 2026-03-06
+-------------------
+
+- :func:`spatialist.ancillary.run`: also return the return code of the subprocess
+
+0.17.1 | 2026-03-10
+-------------------
+
+- :func:`spatialist.ancillary.run`: bug fixes
+
+  + do not encode `inlist` to `bytes`
+  + make sure the function may not return `None`
+  + pas right arguments to `sp.CalledProcessError`
+  + use `with` context manager for the logfile to make sure it is closed